Added font scale support (fscx and fscy override tags) for ASS format subtitles.
Theoretically fixed support for the 'an' (alignment) override tag being combined with a 'pos' position override tag. The 'an' tag was being parsed and handled before the 'pos' override handling which means the 'an' tag wouldn't be present by the time the combined pos/an parsing tried using it. I say theoretically because I don't actually know of a subtitle file that uses this at the moment, but because of the way 'checkOverride' works I know 'an' would have been removed from the overrides before it got to position styling where it tries to read it again.

Fixed/added a few more comments to briefly introduce the quirks of the parser and ASS override tags.