Limit the color blur weakening to just when there's a blur of more than 1px, it's used to soften outline edges at small blur radiuses, and it otherwise makes the outlines kind of disappear in those instances when it's not intended to do so.
Added support for the abbreviated \c override, it's the same as \1c (text fill color)

Added a new parameter to overrides 'repeatable', a boolean that will scan possibly multiple times if an override can be repeated (\t at least, not sure if any others can). I was otherwise seeing an issue where it was failing to match a color it messily matched within a secondary \t override, parsing all of them first prevents conflicts like that.

Only add an alignment after positioning if it's got one, otherwise it'll just do `transform: translate(undefined)` which is messy.