Lots of improvements to parsing. Complex subtitles now scan character by character so it doesn't end on a nested complex override's ending parenthesis (it'll treat the whole nested override as a parameter and not parse its arguments).
Made color parsing more forgiving, it'll wipe out any non-hex characters. I saw some subtitles that had a dangling ')' with no match and it was causing the subtitles to fail to parse because it hadn't properly padded the color string with that extra character in it.

Added a test to test out the parsing of a nested complex override so I know I did it right!