Separated the single subtitle renderer Subtitles.svelte into two different components SubRipRenderer.svelte and ASSRenderer.svelte, each matching their respective parsers. The main motivation behind this was to more easily support some form of collision resolution for ASS subtitles. When two ASS subtitles have the same alignment they should not overlap (except if they're in different layers, not yet supported but could be done fairly easily with this change) and this could be done in a couple different ways.
The first way to support collision resolution, which I didn't choose, is to compute which subtitles are shown at the same time, the size of those subtitles (how many lines they'll turn out to be), and figure out exactly where to absolutely position every line. There's a lot of complexity to it and in the end it'd probably be more accurate, but the complexity was offputting. Especially with the positioning possibly changing if the window is resized (most likely the video is full screened) collision resolution would have to be a renderer concern and couldn't be pre-computed in the parser like everything else has been so far. In the end that in itself isn't a deal breaker, but I decided to go with a simpler solution.

The second way, and the way I actually went with, is to just expose which alignment each subtitle is and mount all alignments of each type within containers positioned to the correct alignment (which is why the ASS specific renderer was needed), that way the subtitles can just flow like normal text within these alignment containers and the browser can handle laying out the text. Text alignment is left/center/right for each container on their respective sides of the screen, and the 1/2/3 alignments (the bottom three) each are `flex-direction: column-reverse` so newer subtitles appear above the most recent one.

There are some caveats with this solution (2/3 can be fairly easily fixed). The margins aren't respected, because they're a little more complicated than they sound (they're margins specifically from the edge of the screen, not just regular css margins, so treating them as regular margins would put them in the wrong spot- i.e. if the subtitle is greater than the margin from the side of the screen it should not move, but with a regular css margin it would move). Layers aren't currently supported, but just duplicating the alignment container structure for each layer at play would be pretty easy to implement. There's also another scenario where two subtitles are showing at once in the same alignment, the first one goes away and a third one shows up, in this scenario subtitle three should replace the first subtitle, so subtitle 2 doesn't randomly jump around. This isn't supported yet (subtitle two would shift to where one was, and three would take two's place, making it really hard to read text that's moving around), but could be done somewhat easily by looping through subtitles after parsing and giving each subtitle that has an alignment a value that will be used as a flexbox 'order' property that's the first available index when a subtitle starts. In the previous scenario subtitle one would be given order: 1, subtitle two would be given order: 2, then when subtitle one goes away and the third one shows up it'd also get order: 1, because only "order: 2" is spoken for and 1 is available.

Now that the non-positioned alignment is handled strictly by whichever container the subtitle is mounted in, positioned subtitles aren't impacted by their inherited style's alignment transform. This means positioned subtitles won't move

Also added support for the legacy \a alignment override, some tweaks to perspective to make 3d rotation a tiny bit more accurate, fixed text info parsing when there's no space between the info block's ':' and the value (undefined playResX/playResY makes any scaled positioning/font size get NaN), use a drop-shadow filter for the text shadows which looks a lot better.

Also renamed "SRT" classes to "SubRip" as it's the actual name of the format, not just the file extension, looked better!