Added an initial implementation of drawing commands. They're not perfect and don't scale properly with the player so they likely won't look the right size unless you're at a similar resolution to the script's playres. Some later (possibly runtime based) scaling will probably have to be done form the renderer. For the time being though the drawings generate an SVG path with a little bit of tweaking to the draw commands, they're fairly similar thankfully. That SVG just gets mounted as straight HTML by the renderer.
Combined movement tween stores because svelte can support tweening between objects of the same shape, makes the that movement animation look a lot cleaner code-wise.

The debug output for ASS subtitles at a specific time will now filter styles down to only the ones relevant to the displayed lines. Some scripts have tens of styles and it's a lot to look through if it doesn't actually matter to you.

Found in the ASS override tag docs that movement timings of [0, 0] is equivalent to not having timings, so added a case for that.