Rewrote the player to stream DASH videos.
To accomplish the goal of being able to switch between multiple audio streams the previous solution was to mux multiple copies of the video of each audio track combination. This was inefficient, it required twice the storage space and switching streams would reset the subtitler as it'd change the `src` on the video.

This is step one of switching to a better pattern. Media Source Extensions let audio/video be loaded into buffers at will, so the audio track switching goal can be accomplished by removing and loading a different audio file's buffer into the media source. This requires the video buffering to be completely controlled by the application. The videos are almost certainly too big to load all at once so they get DASHed by jimaku-player-ingest and the resulting MPD file is embedded in the video metadata, then parsed by MPDParser and buffered by Streamer and Saisei. Getting things to load properly is pretty difficult so a logger has been made that logs things about what segments get buffered and what's expected to be buffered. It's currently enabled but will be disabled eventually as the rest of the features are sorted out.