Separate audio and video streams, they're now dashed independently. This includes inspecting the video and audio buffer health independently and moving the responsibility for that to Streamer instead of Saisei where it actually has access to both buffers. Because each audio/video stream is dashed independently MPDParser was rewritten to only expect a single AdaptationSet for a single stream of media, so no video/audio difference there as that's handled by Streamer. Saisei once again can switch audio tracks and that's done by swapping out the audio resource Streamer uses to fill the audio buffer.
Also disabled the Svelte caption warning on video elements as the whole point of this project is to provide subtitles, and disabled selecting the "Paused" text when the video is paused.

The health of the video is now checked only once per second instead of whenever `currentTime` changes. This was necessary because if you seek through the video by dragging the range input it will try and load every single segment along the way which could be hundreds of megabytes of segments that immediately get thrown away.