Better vertical alignment ('line' cue) support for WebVTT subtitles, and made the clickable area for a Jisho lookup more closely match the width of the text so you're less likely to accidentally look something up when you were intending on clicking on something else.
Rewrote a bit of how SubRip/WebVTT subtitles are rendered. The renderer still had some remaining artifacts from when it was shared with ASS subtitles. Since it has no need for distinct phrases the structure was cleaned up a bit and now there's just a single 'p' element instead of it containing a 'span' inside with distinct styles on both which makes everything a bit easier to deal with. Some basic styles that don't change were moved out of the parser's inline styles and into the renderer's component classes.

Added better support for WebVTT 'line' arguments. There are some special cases, like '0' and '-1' as non-percentages, that mirror '0%' and '100%' respectively. I don't think I'm exactly handling this as it seems it can be positive or negative integers judging by what MDN says (not just 0 and -1) but this anyways seems to work for those two currently anyway and it can be re-evaluated in the future.

Added a bit of code to swap between two different clamped values for the 'top' on a subtitle, this way if the subtitles are showing on the bottom of the screen they hopefully won't overlap the video controls when we expect them to show up (user is active).

The 'userActive' store now is better at telling when the user is actually active. If you kept moving the mouse it'd timeout then immediately reset back to active leading to things kinda jumping around for a frame, it just wasn't properly extending the timeout like I had intended.

Since there was no good format test script I could find like Aegisub's ass-format-tests.ass I started my own demo vtt script, and made a small test app that shows a native browser rendering (using a 'track' element) and the SubRipRenderer.svelte verison side-by-side for every subtitle line in the script so you can just scroll down the page and see if everything looks right. Some things that were reused (the header and the GM_* shims) needed to be broken out into their own files.