Small bug fix release to fix lowercase ASS colors.
The color parsing in ASS subtitles didn't work well with lowercase hex characters so a color that should be white like &Hffffff would get stripped to &H and padded with 0's, which was intended to only pad the alpha channel because it's often omitted but in this case was turning the whole color black. This fix just upper cases the color strings before doing anything else with them.