Provide fallbacks for playResX and playResY
For some reason a subtitle file might not have `playResX` or `playResY` (video player resolutions) in the subtitle file's info block. Without these numbers anything that scales to the size of the player, specifically font sizes will result in `NaN` which means the subtitle text will be really small and it won't be able to be scaled via the font scale settings either as `NaN * 3 = NaN` still.

This just adds some default sizes if they weren't provided by the subtitle script. They might not match the video resolution, but fonts can be scaled with that option still.

Closes #33