Updated the Japanese font stack to move Noto Sans to after its JP and CJK variants.
I recently switched to Fedora and with this font stack I was seeing the Chinese variants of characters. My best guess is because I have Noto Sans but that doesn't have Japanese glyphs it falls back to the `sans-serif` fallback, skipping the more specific variants that do have those glyphs. Moving the normal Noto Sans to after the rest of those seems to fix my issue.