Typescript/Vite rewrite and refactoring
There was a lot of stuff that was all somewhat related so I did it all at once.

I wanted to rewrite all of the front end code (the subtitler itself and the dev site) in Typescript, to do that I wanted to also switch to using Vite instead of webpack. There were a few things here and there that needed to change about how some things were done (particularly the subtitle parsers) to fit with Typescript, but all in all it's roughly the same.

I also did a bit of an update to the style used by the subtitler itself. It has it's own styling that matches the other sites I made that use sheodox-ui instead of trying to mimic the VRV style with semi-transparent blue/purplish backgrounds and yellow buttons. Again it's mostly the same, but just looks a little better.

One bigger change is the addition of an About modal which now holds credits, links, and the debug information instead of hiding it all in a collapsible thing at the bottom of the Settings tab.

I removed the dependency on Postgres/TypeORM for the dev server. It was only really used for storing thumbnails which are now just built and put on the file system. Each video (well each path to a video) gets an ID which matches it to a thumbnail. There's now support again for regular mp4 videos too, as I rewrote all of the video scanning code.

I also added some steps to lint code on pre-commit.