Constrain subtitles to an aspect ratio
Until now the whole viewport (of the video player iframe) was treated as the video, meaning subtitles that were really long, or anchored to a side or corner would not be over the video where it "should" be unless you had a monitor that matched the aspect ratio, or you resized the window weirdly.

To fix this theres a new "Aspect Ratio" setting, which gives you the choice between 16:9, 4:3, and 'auto'. For ASS subtitles 'auto' will use the PlayResX and PlayResY metadata to guess the aspect ratio, otherwise it'll use 16:9, but the other 16:9 and 4:3 settings will force an aspect ratio regardless of what the subtitles say.

This generally was done by setting a width on the container of the subtitle renderers with the aspect ratio times 100vh and center it, creating a container that should match exactly where the video sits. The `left` on all of the ASS alignments also takes this aspect ratio into account.

Part of why this wasn't done earlier was that setting a container to take up the full viewport width means there's an invisible div covering all of the controls and blocking your clicks, but I got around this with a combination of `pointer-events: none` and `pointer-events: auto` on descendent elements.

Closes #34