Centered the tray vertically.
The tray was previously close to the top, but depending on your setup there might be buttons up there, to further get out of the way I've moved the tray down a bit further by centering it, and removing the top margin. Just so this hopefully doesn't overlap buttons on the bottom again I made the tray a bit shorter too.