Fixed VTT positioning and cleaned up subtitle text in some cases
I was using some transforms to account for SRT not having `line` properties, but that had an unintended side effect of not trusting VTT to specify appropriate `line` values and the transformY that was applied pushed the subtitles too close to the center of the video.

Also removing some html entities (more can be added later). Someone encountered `&lrm;` which isn't a visible character and it looked fine without it. Normally svelte will replace html entities but not in dynamic expressions. For this case I decided to just remove the entity.

I also made the cue payload text tags removal work with more usages and not just specifically 'c.Japanese'.