Added hotkeys for fine alignment adjustment without having to open the Fine Adjustment dialog
Pressing `z` or `x` will delay or hasten the current alignment (in place, over the top of the most recent thing in the history) by 0.2s or by 1s if shift is held.

I had to update some of the toast code to allow a specific ID which lets you update a toast instead of just adding a new one so the text of the current toast gets updated instead of spamming toasts when it displays the new alignment.