Hotkey to go back to rewind to last sub, style updates and fixes
There's a lot in this, updated a bunch of dependencies, including sheodox-ui for updated base styles, and I improved the layout and wording of a lot of stuff in the tray so it's less of a soup of confusing buttons.

I added a hotkey 'b' which rewinds just before the last sub, you can hit this multiple times to keep going back line by line. This required a rewrite of how sub history is decided and it now shows subs from before the most recent subtitles instead of manually manipulating an array with any subtitles we haven't seen before every time they change. This will make the recent subs tab accurate to that point in time when you rewind or skip through a video instead of it being a mismash of subtitles from everywhere.

I also fixed some styling issues, Crunchyroll defined .row to be a flex row with justify-content: space-between, and I was using it to be just a row in a column, nothing with its actual display. I used f-row where appropriate, but it could probably use some cleanup where it's not necessary anymore.